refactor: `ScriptResource` can now only be created using factory methods
This does make the class much more complicated again, but now the code
is much more directly connected to its actual purpose.
